feat!: encupslate all S3 opeartions in one class for easier maintainability
