--- utils/storage/S3.js
@@ -1,13 +1,0 @@
-const { S3Client } = require('@aws-sdk/client-s3')
-require('dotenv').config()
-const process = require('process')
-
-const S3 = new S3Client({
-    region: process.env.AWS_REGION,
-    credentials: {
-        accessKeyId: process.env.AWS_ACCESS_KEY_ID,
-        secretAccessKey: process.env.AWS_SECRET_ACCESS_KEY,
-    },
-})
-
-module.exports = S3
