--- utils/storage/bufferToStream.js
@@ -1,10 +1,0 @@
-const { Readable } = require('node:stream')
-
-function bufferToStream(buffer) {
-    const stream = new Readable()
-    stream.push(buffer)
-    stream.push(null)
-    return stream
-}
-
-module.exports = bufferToStream
